DOCS-2234 Fusion 5 Upgrade Docs
Currently, we include a short snippet for upgrading Fusion 5 which is catered towards users on GKE. For the release of Fusion 5.1, we should include links to the upgrade instructions for each deployment option. This PR tags the upgrade instructions for EKS, AKS, GKE, and other K8s deployments. This will allow us to include the information in the docs. (DOCS-2234)