APOLLO-26331 Include Seldon deployments in scale-down
This adds 'scaling-down' of Seldon deployments. This is a lie. It actually deletes all Seldon deployments within a namespace as `kubectl scale` is not apparently supported with the version of Seldon we're deploying in Fusion 5.0-5.2. When we upgrade to 1.2.x in Fusion 5.3, scaling should be supported, so I've included the eventual `kubectl scale` command as a comment for us to come back to later.